# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "twp server api.",
        "title": "twp API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.o"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/user": {
            "get": {
                "description": "Get all user information. Include user's icon, name, email, created time and role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "User"
                ],
                "summary": "Admin Get User",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Begin index",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "maximum": 20,
                        "type": "integer",
                        "default": 10,
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.GetUsersRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/admin/user/{username}": {
            "delete": {
                "description": "Disable user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "User"
                ],
                "summary": "Admin Disable User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/buyer/cart": {
            "get": {
                "description": "Get all Carts of the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buyer",
                    "Cart"
                ],
                "summary": "Buyer Get Cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/buyer.Cart"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/buyer/cart/{cart_id}/checkout": {
            "get": {
                "description": "Get all checkout data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buyer",
                    "Checkout"
                ],
                "summary": "Buyer Get Checkout",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cart ID",
                        "name": "cart_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/buyer.checkout"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Checkout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Buyer",
                    "Checkout"
                ],
                "summary": "Buyer Checkout",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cart ID",
                        "name": "cart_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment",
                        "name": "payment_method",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/buyer.PaymentMethod"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "buyer.Cart": {
            "type": "object",
            "properties": {
                "cartInfo": {
                    "$ref": "#/definitions/db.GetCartRow"
                },
                "coupons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.GetCouponsFromCartRow"
                    }
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.GetProductFromCartRow"
                    }
                }
            }
        },
...
        "buyer.PaymentMethod": {
            "type": "object",
            "properties": {
                "creditCard": {
                    "type": "object"
                }
            }
        },
...
        "buyer.checkout": {
            "type": "object",
            "properties": {
                "coupons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/buyer.couponDiscount"
                    }
                },
                "payments": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "shipment": {
                    "type": "integer"
                },
                "subtotal": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_discount": {
                    "type": "integer"
                }
            }
        },
        "buyer.couponDiscount": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "discount_value": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/db.CouponScope"
                },
                "type": {
                    "$ref": "#/definitions/db.CouponType"
                }
            }
        },
...
        "db.CouponScope": {
            "type": "string",
            "enum": [
                "global",
                "shop"
            ],
            "x-enum-varnames": [
                "CouponScopeGlobal",
                "CouponScopeShop"
            ]
        },
...
        "db.CouponType": {
            "type": "string",
            "enum": [
                "percentage",
                "fixed",
                "shipping"
            ],
            "x-enum-varnames": [
                "CouponTypePercentage",
                "CouponTypeFixed",
                "CouponTypeShipping"
            ]
        },
...
        "db.GetCartRow": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "seller_name": {
                    "type": "string"
                },
                "shop_image_url": {
                    "type": "string"
                },
                "shop_name": {
                    "type": "string"
                }
            }
        },
        "db.GetCouponsFromCartRow": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/db.CouponScope"
                },
                "type": {
                    "$ref": "#/definitions/db.CouponType"
                }
            }
        },
...
        "db.GetProductFromCartRow": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "db.GetUsersRow": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "credit_card": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "email": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "icon_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/db.RoleType"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "db.RoleType": {
            "type": "string",
            "enum": [
                "admin",
                "customer"
            ],
            "x-enum-varnames": [
                "RoleTypeAdmin",
                "RoleTypeCustomer"
            ]
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}